Replace Phone class with Priority class
